refactor: 공개프로필 훅, 컴포넌트 분리
